Reorganize the components folder
